% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_effsize.R
\name{compute_effsize}
\alias{compute_effsize}
\title{Calculates the effect size given two prob distributions on finite set}
\usage{
compute_effsize(p1, p2)
}
\arguments{
\item{p1}{A vector of probs.}

\item{p2}{A second vector of probs.}
}
\value{
The Vargha and Delaney's A effect size of a population with probability distributions given by p1 and p2.
}
\description{
Calculates the effect size given two prob distributions on finite set
}
\examples{
compute_effsize(c(.5, .5), c(.4, .6)
}
