% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_power.R
\name{estimate_power}
\alias{estimate_power}
\title{Estimates power given a method of sampling from two distributions}
\usage{
estimate_power(
  sample_size,
  dsamp1,
  dsamp2,
  alpha = 0.05,
  any_ties = TRUE,
  num_replicate = 10000
)
}
\arguments{
\item{sample_size}{Number of samples in each population}

\item{dsamp1}{Method of sampling from population 1}

\item{dsamp2}{Method of sampling from population 2}

\item{alpha}{Significance level of the hypothesis test}

\item{any_ties}{Set to FALSE if there are no possible ties in the populations}

\item{num_replicate}{Number of replications in simulated power. Default 10000}
}
\value{
Estimate of power from wilcox.test
}
\description{
Estimates power given a method of sampling from two distributions
}
\examples{
estimate_power(sample_size = 40)

}
